+++ plato_calendar/lib/Page/settings.dart
@@ -12,8 +12,8 @@ import '../Data/ics.dart';
 import '../utility.dart';
 import 'widget/appointmentEditor.dart';
 import '../main.dart';
-import '../plato.dart';
-
+import '../pnu/plato.dart';
+import '../pnu/pnu.dart';
 
 class Setting extends StatefulWidget{
   @override
@@ -353,7 +353,7 @@ class _LoginPageState extends State<LoginPage> {
                 });
                 UserData.id = idController.text;
                 UserData.pw = pwController.text;
-                await Plato.update(force: true);
+                await update(force: true);
                 Navigator.pop(context, true);
               },
               child: Container(
@@ -396,7 +396,7 @@ class Loading extends AnimatedWidget{
                 if(!controller.isAnimating)
                   if(DateTime.now().difference(_manualUpdateTime).inMinutes > 4){
                     controller.repeat();
-                    await Plato.update(force : true).then((value) {
+                    await update(force : true).then((value) {
                         if(value) _manualUpdateTime = DateTime.now();
                     });
                     controller.stop();
